[PATCH] USB: PL2303 - PharosGPS patch
This patch adds recognition/support for the PharosGPS puck, as included with
Microsoft's Streets and Trips package.  The device has an inline
USB-to-serial converter which functions as a PL2303 but with different
VENDOR_ID and PRODUCT_ID, and so the PL2303 driver is modified to recognize
this additional device.

Signed-off by: Robert C. Olsen, III [pl2303@2005dauphin.org]
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>