[PATCH] boot parameters: quoting of environment variables revisited
As noticed by Joey Hess (and thanks for Christoph for forwarding it).  Also
requirements from Werner Almesberger.

If someone passes 'foo="some value"' the param engine removes the quotes
and hands 'foo' and 'some value'.  The __setup() parameters expect a single
string, and so we try to regenerate it from the two parts.  Finally, we try
to place it as an environment variable.

Werner wants quotes stripped out of the environment variable.  It makes
sense to do that for __setup, too (so it sees 'foo=some value'), since
__setup functions don't usually handle quotes.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>